+++ src/main/kotlin/com/project/goms/domain/account/presentation/data/response/ProfileResponse.kt
@@ -6,7 +6,7 @@ data class ProfileResponse(
     val accountIdx: UUID,
     val studentNum: StudentNum,
     val profileUrl: String?,
-    val rateCount: Int
+    val rateCount: Long
 ) {
     data class StudentNum(
         val grade: Int,
